feat: added navigation from NearbyInterests to its detail page (#151)
