+++ routes/routeStack.js
@@ -7,7 +7,7 @@ import MapDirections from "../screens/Directions";
 import PreviewDirections from "../screens/PreviewDirections";
 import PreferenceMenu from "../screens/PreferenceMenu";
 import IndoorMapView from "../screens/Indoor/IndoorMapView";
-import Nearby from "../screens/Menu/Nearby";
+import Nearby from "../screens/Menu/NearbyInterest";
 
 // Add screens here
 const screens = {
